Added motif docstring | refactored tests
